feat: added user service implementations for no-ldap profile
