+++ digiwf-engine/digiwf-engine-service/src/main/java/io/muenchendigital/digiwf/legacy/user/domain/service/UserService.java
@@ -1,30 +1,11 @@
-/*
- * Copyright (c): it@M - Dienstleister für Informations- und Telekommunikationstechnik der Landeshauptstadt München, 2020
- */
-
 package io.muenchendigital.digiwf.legacy.user.domain.service;
 
 import io.muenchendigital.digiwf.legacy.user.domain.model.User;
-import io.muenchendigital.digiwf.legacy.user.infrastructure.UserRepository;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.apache.commons.lang3.StringUtils;
-import org.springframework.stereotype.Service;
 
-import java.util.*;
-import java.util.stream.Collectors;
+import java.util.List;
+import java.util.Optional;
 
-/**
- * Service to query for users.
- *
- * @author externer.dl.horn
- */
-@Slf4j
-@Service
-@RequiredArgsConstructor
-public class UserService {
-
-    private final UserRepository userRepository;
+public interface UserService {
 
     /**
      * Get groups for the given user id.
@@ -32,9 +13,7 @@ public class UserService {
      * @param userId Id for the user
      * @return groups in lowercase
      */
-    public List<String> getGroups(final String userId) {
-        return this.userRepository.findOuTree(userId).stream().map(String::toLowerCase).collect(Collectors.toList());
-    }
+    List<String> getGroups(final String userId);
 
     /**
      * Search users by the given string and groups.
@@ -43,28 +22,7 @@ public class UserService {
      * @param ous          Groups of the user
      * @return users
      */
-    public List<User> searchUser(final String searchString, final String ous) {
-        log.debug("Searching user with {}", searchString);
-
-        if (StringUtils.isEmpty(searchString)) {
-            throw new IllegalArgumentException("Expected value in request: queryString");
-        }
-
-        final List<String> oufilterList = ous == null ? Collections.emptyList() : Arrays.asList(ous.split(","));
-
-        final String[] queryStrings = searchString.strip().split(" ");
-        final List<User> userInfos = new ArrayList<>();
-        if (queryStrings.length >= 2 && !StringUtils.isEmpty(queryStrings[0]) && !StringUtils.isEmpty(queryStrings[1])) {
-            final List<User> userInfosByName = this.userRepository.findByNamesLike(queryStrings[0].strip(), queryStrings[1].strip(), oufilterList);
-            userInfos.addAll(userInfosByName);
-        } else {
-            final List<User> userInfosByName = this.userRepository.findByNamesLike(searchString.strip(), oufilterList);
-            userInfos.addAll(userInfosByName);
-        }
-
-        log.debug("Hits for {}: {}", searchString, userInfos.size());
-        return userInfos;
-    }
+    List<User> searchUser(final String searchString, final String ous);
 
     /**
      * Format a specific user.
@@ -72,18 +30,7 @@ public class UserService {
      * @param userId Id of the user
      * @return the formatted user string
      */
-    public String getUserString(final String userId) {
-        if (StringUtils.isBlank(userId)) {
-            return null;
-        }
-        final Optional<User> user = this.getUserOrNull(userId);
-        return user.map(value -> value.getForename() +
-                " " +
-                value.getSurname() +
-                " (" +
-                value.getOu() +
-                ")").orElse(userId);
-    }
+    String getUserString(final String userId);
 
     /**
      * Get user by id.
@@ -91,10 +38,7 @@ public class UserService {
      * @param userId Id of the user
      * @return user
      */
-    public User getUser(final String userId) {
-        return this.userRepository.findById(userId)
-                .orElseThrow(() -> new IllegalArgumentException(String.format("User with the id %s does not exist.", userId)));
-    }
+    User getUser(final String userId);
 
     /**
      * Get user by id.
@@ -102,9 +46,7 @@ public class UserService {
      * @param userId Id of the user
      * @return user optional
      */
-    public Optional<User> getUserOrNull(final String userId) {
-        return this.userRepository.findById(userId);
-    }
+    Optional<User> getUserOrNull(final String userId);
 
     /**
      * Get user by username.
@@ -112,15 +54,12 @@ public class UserService {
      * @param username Username of the user
      * @return user optional
      */
-    public Optional<User> getUserByUserName(final String username) {
-        return this.userRepository.findByUsername(username);
-    }
+    Optional<User> getUserByUserName(final String username);
 
     /**
      * Get OU by shortname.
      */
-    public Optional<User> getOuByShortName(final String shortname) {
-        return this.userRepository.findOuByShortName(shortname);
-    }
+    Optional<User> getOuByShortName(final String shortname);
+
 
 }
